consolidate logging utilities
