+++ src/automation/automator.py
@@ -2,7 +2,7 @@ import time
 
 from src.vision.ocr import capture_screen, extract_text
 from src.vision.states import detect_state, handle_state
-from src.utils.logger import save_screenshot, log_ocr_text
+from utils.logger import save_screenshot, log_ocr_text
 from . import mode_manager
 from .quest_path import visit_trainer_if_needed
