+++ tests/test_utils_logger.py
@@ -8,7 +8,7 @@ from importlib import reload
 def test_save_screenshot_creates_file(tmp_path, monkeypatch):
     fake_cv2 = types.SimpleNamespace(imwrite=lambda p, img: open(p, "wb").close())
     monkeypatch.setitem(sys.modules, "cv2", fake_cv2)
-    import src.utils.logger as logger
+    import utils.logger as logger
     reload(logger)
 
     out_dir = tmp_path / "shots"
@@ -17,7 +17,7 @@ def test_save_screenshot_creates_file(tmp_path, monkeypatch):
 
 
 def test_log_ocr_text(tmp_path):
-    from src.utils import logger
+    from utils import logger
     log_file = tmp_path / "ocr.log"
     logger.log_ocr_text("hello", log_path=str(log_file))
     content = log_file.read_text()
@@ -25,7 +25,7 @@ def test_log_ocr_text(tmp_path):
 
 
 def test_log_event(tmp_path):
-    from src.utils import logger
+    from utils import logger
     log_file = tmp_path / "ev.log"
     logger.log_event("hi", log_path=str(log_file))
     assert log_file.exists()
@@ -34,7 +34,7 @@ def test_log_event(tmp_path):
 
 
 def test_log_performance_summary(tmp_path):
-    from src.utils import logger
+    from utils import logger
 
     csv_file = tmp_path / "perf.csv"
     stats = {"quests": 1, "xp": 42}
